Module Analyses.Disjoint

module VSet : sig ... end

module VMap : sig ... end

include sig ... end

type t = VSet.t VMap.t Soteria_std.Dynarray.t

val init : unit -> t

val backtrack_n : t -> int -> unit

val save : t -> unit

val reset : t -> unit

val set_default : VSet.t VMap.t -> unit

val wrap : (VSet.t VMap.t -> 'a * VSet.t VMap.t) -> t -> 'a

val wrap_read : (VSet.t VMap.t -> 'a) -> t -> 'a

val pp : Stdlib.Format.formatter -> VSet.t VMap.t -> unit

val sort : Svalue.t -> Svalue.t -> Svalue.t * Svalue.t

val var_set : Svalue.t -> Soteria.Symex.Var.Set.t

val is_relevant : Svalue.t -> bool

val ineq_matters : Svalue.t -> bool

val add_inequalities : VMap.key list -> VSet.t VMap.t -> Soteria.Symex.Var.Set.t * VSet.t VMap.t

Adds inequalities between all values in vs to the state st, returning the updated state along with the set of variables that are dirty (i.e. something was learnt from this).

val sure_neq : Svalue.t -> Svalue.t -> VSet.t VMap.t -> bool

val simplify : t -> Svalue.t -> Soteria__Bv_values__Svalue.t_node Hc.hash_consed

val add_constraint : t -> Svalue.t -> Soteria__Bv_values__Svalue.t_node Hc.hash_consed * Soteria.Symex.Var.Set.t

val filter : t -> Soteria.Symex.Var.t -> Svalue.ty -> Svalue.t Soteria_std.Iter.t -> Svalue.t Soteria_std.Iter.t

val mk_var : int -> Soteria.Symex.Var.t -> [> Typed.T.sint ] Typed.t

val uneq : Svalue.t -> Svalue.t -> [> Typed.sbool ] Typed.t

val encode : ?vars:Soteria.Symex.Var.Hashset.t -> t -> Typed.sbool Typed.t Soteria_std.Iter.t