Parameter Merge.A2

include Soteria.Soteria_std.Reversible.Mutable

type t

val init : unit -> t

Create a new reversible state initialized with the default value.

val backtrack_n : t -> int -> unit

Remove the last n checkpoints from state.

val save : t -> unit

Save the current state as a new checkpoint.

val reset : t -> unit

Clear all checkpoints and reset to the default value.

val simplify : t -> Svalue.t -> Svalue.t

Simplifies a constraints using the current knowledge base, without updating it.

val add_constraint : t -> Svalue.t -> Svalue.t * Soteria.Symex.Var.Set.t

Adds a constraint to the current analysis, updating the currently tracked data.

val filter : t -> Soteria.Symex.Var.t -> Svalue.ty -> Svalue.t Soteria_std.Iter.t -> Svalue.t Soteria_std.Iter.t

Filters the given iterator of symbolic values, keeping only those relevant to the given variable according to the analysis.

val encode : ?vars:Soteria.Symex.Var.Hashset.t -> t -> Typed.sbool Typed.t Soteria_std.Iter.t

Encode all the information relevant to the given variables and conjuncts them with the given accumulator.