Module Encoding.LetBinder

module ValTbl : sig ... end

module VarSet : sig ... end

val var_of_value : Svalue.t -> Soteria.Symex.Var.t

Generates an appropriate Var that uniquely identifies this value.

val val_of_value : Svalue.t -> Svalue.t

Like var_of_value but returns a properly typed Svalue.

val dependencies : Svalue.t -> VarSet.t

Returns the set of Vars mentionned in this value.

val replace : ?cond:(Svalue.t -> bool) -> Svalue.t -> Svalue.t

replace ?cond v replaces v with a Var node with id v.tag if cond v is true (cond defaults to Fun.const true); otherwise recursively replaces the children.

val mk_bindings : 'a ValTbl.t -> (Soteria.Symex.Var.t * Simple_smt.sexp * VarSet.t) list

Makes the bindings for a map of bindings (ignores the pointed-to value). Will return, for each binding, an appropriate variable for the binding (using var_of_value), the encoded value (with possible sub-bindings replaced), and a set containing the binding dependencies of this value.

val order_bindings : VarSet.t -> (Soteria.Symex.Var.t * Simple_smt.sexp * VarSet.t) list -> (string * Simple_smt.sexp) list list

order_bindings known_vars bindings will order the given bindings into groups of bindings; each group only needs the variables in the previous groups or in known_vars to be defined.

val let_binds_for : ?min_binds:int -> ?min_occurrences:int -> Svalue.t -> Svalue.t * (string * Simple_smt.sexp) list list

Will calculate the let-bindings for the given value. Will only consider bindings for values that occur at least min_occurrences times, and will only return any bindings at all if at least min_binds are found matching the above criterium. Will return the value with any relevant binding substituted (but not defined in the AST!) along with the bindings to be applied after encoding.

val apply_bindings : (string * Simple_smt.sexp) list list -> Simple_smt.sexp -> Simple_smt.sexp

apply_bindings bindings sexp Applies the given groups of bindings to sexp.