Module Typed.T

type sint = [

| `NonZero

| `Zero

]

A symbolic integer; can either be `NonZero if it is known to not be 0, `Zero if it is 0.

type sint_ovf = [

| `NonZero

| `Zero

| `Overflowed

]

Any symbolic integer; it may be the result of an overflowing operation

type nonzero = [

| `NonZero

]

A symbolic integer known to be non-zero.

type zero = [

| `Zero

]

A symbolic integer known to be zero.

type sfloat = [

| `Float

]

type sbool = [

| `Bool

]

type sptr = [

| `Ptr

]

type sloc = [

| `Loc

]

type 'a sseq = [

| `List of 'a

]

type cval = [

| sint

| sptr

| sfloat

]

type any = [

| sint_ovf

| sfloat

| sbool

| sptr

| sloc

| any sseq

]

val pp_sint : Stdlib.Format.formatter -> sint -> unit

val pp_sint_ovf : Stdlib.Format.formatter -> sint_ovf -> unit

val pp_nonzero : Stdlib.Format.formatter -> nonzero -> unit

val pp_zero : Stdlib.Format.formatter -> zero -> unit

val pp_sfloat : Stdlib.Format.formatter -> sfloat -> unit

val pp_sbool : Stdlib.Format.formatter -> sbool -> unit

val pp_sptr : Stdlib.Format.formatter -> sptr -> unit

val pp_sloc : Stdlib.Format.formatter -> sloc -> unit

val pp_cval : Stdlib.Format.formatter -> cval -> unit

val pp_sseq : (Stdlib.Format.formatter -> 'a -> unit) -> Stdlib.Format.formatter -> 'a sseq -> unit

val pp_any : Stdlib.Format.formatter -> any -> unit

val hash_sint : sint -> int

val hash_sint_ovf : sint_ovf -> int

val hash_nonzero : nonzero -> int

val hash_zero : zero -> int

val hash_sfloat : sfloat -> int

val hash_sbool : sbool -> int

val hash_sptr : sptr -> int

val hash_sloc : sloc -> int

val hash_cval : cval -> int

val hash_sseq : 'a sseq -> int

val hash_any : any -> int