Module C_values.C_solver

module Z3_solver : Soteria_symex.Solver.Mutable_incremental with module Value = Typed

This module represents a solver state, it is fully imperative!

module Z3_incremental_solver : Soteria_symex.Solver.Mutable_incremental with module Value = Typed

This module represents a solver state, it is fully imperative!