Module C_values.Svalue

module Var = Soteria.Symex.Var

module FloatPrecision : sig ... end

module FloatClass : sig ... end

module RoundingMode : sig ... end

type ty =

| TBool

| TInt

| TFloat of FloatPrecision.t

| TLoc

| TPointer

| TSeq of ty

| TBitVector of int

val equal_ty : ty -> ty -> Ppx_deriving_runtime.bool

val pp_ty : Ppx_deriving_runtime.Format.formatter -> ty -> Ppx_deriving_runtime.unit

val show_ty : ty -> Ppx_deriving_runtime.string

val compare_ty : ty -> ty -> Ppx_deriving_runtime.int

val t_bool : ty

val t_int : ty

val t_float : FloatPrecision.t -> ty

val t_f16 : ty

val t_f32 : ty

val t_f64 : ty

val t_f128 : ty

val t_loc : ty

val t_ptr : ty

val t_seq : ty -> ty

val t_bv : int -> ty

val is_float : ty -> bool

val precision_of_f : ty -> FloatPrecision.t

val size_of_bv : ty -> int

module Nop : sig ... end

module Unop : sig ... end

module Binop : sig ... end

val pp_hash_consed : ('a -> 'b -> 'c) -> 'a -> 'b Hc.hash_consed -> 'c

val equal_hash_consed : 'a -> 'b Hc.hash_consed -> 'c Hc.hash_consed -> bool

val compare_hash_consed : 'a -> 'b Hc.hash_consed -> 'c Hc.hash_consed -> int

type t_kind =

| Var of Var.t

| Bool of bool

| Int of Z.t

| Float of string

| Ptr of t * t

| BitVec of Z.t

| Seq of t list

| Unop of Unop.t * t

| Binop of Binop.t * t * t

| Nop of Nop.t * t list

| Ite of t * t * t

and t_node = {

kind : t_kind;

ty : ty;

}

and t = t_node Hc.hash_consed

val pp_t_kind : Ppx_deriving_runtime.Format.formatter -> t_kind -> Ppx_deriving_runtime.unit

val show_t_kind : t_kind -> Ppx_deriving_runtime.string

val pp_t_node : Ppx_deriving_runtime.Format.formatter -> t_node -> Ppx_deriving_runtime.unit

val show_t_node : t_node -> Ppx_deriving_runtime.string

val show : t -> Ppx_deriving_runtime.string

val equal_t_kind : t_kind -> t_kind -> Ppx_deriving_runtime.bool

val equal_t_node : t_node -> t_node -> Ppx_deriving_runtime.bool

val compare_t_kind : t_kind -> t_kind -> Ppx_deriving_runtime.int

val compare_t_node : t_node -> t_node -> Ppx_deriving_runtime.int

val hash : 'a Hc.hash_consed -> int

val kind : t_node Hc.hash_consed -> t_kind

val iter_vars : t -> ((Var.t * ty) -> unit) -> unit

val pp_full : Ppx_deriving_runtime.Format.formatter -> t_node Hc.hash_consed -> Ppx_deriving_runtime.unit

val pp : t Fmt.t

val equal : 'a Hc.hash_consed -> 'b Hc.hash_consed -> bool

val compare : 'a Hc.hash_consed -> 'b Hc.hash_consed -> int

val sure_neq : t -> t -> bool

module Hcons : sig ... end

val (<|) : t_kind -> ty -> t

val mk_var : Var.t -> ty -> t

val mk_commut_binop : Binop.t -> t -> t -> t_kind

We put commutative binary operators in some sort of normal form where element with the smallest id is on the LHS, to increase cache hits.

val subst : (Var.t -> Var.t) -> t -> t

Booleans

val v_true : t

val v_false : t

val as_bool : 'a Hc.hash_consed -> bool option

val bool : bool -> t

val and_ : t -> t -> t

val or_ : t -> t -> t

val conj : t list -> t

val not : t -> t

val split_ands : t -> (t -> unit) -> unit

val distinct : t list -> t

val ite : t -> t -> t -> t

Integers

val int_z : Z.t -> t

val int : int -> t

val nonzero_z : Z.t -> t

val nonzero : int -> t

val zero : t

val one : t

val int_of_bool : t -> t

val plus : t -> t -> t

val minus : t -> t -> t

val times : t -> t -> t

val div : t -> t -> t

val is_mod : t -> Z.t -> bool

val rem : t -> t -> t

val mod_ : t -> t -> t

val neg : t -> t

module BitVec : sig ... end

module Float : sig ... end

val lt : t -> t -> t

val leq : t -> t -> t

val geq : t -> t -> t

val gt : t -> t -> t

val sem_eq : t -> t -> t

val sem_eq_untyped : t -> t -> t

val not_int_bool : t -> t

Negates a boolean that is in integer form (i.e. 0 for false, anything else is true)

val bool_of_int : t -> t

Pointers

module Ptr : sig ... end

Sequences

module SSeq : sig ... end

Infix operators

module Infix : sig ... end

module Syntax : sig ... end