Module Svalue.BitVec

Bit vectors

Bit vector operations; all these expect values of type TInt, and apply bit-vector conversions under the hood as needed. This allow bit operations to be seamlessly used on TInt, without worrying about the underlying representation.

val bool_or : t -> t -> t

val bool_and : t -> t -> t

val bool_not : t -> t

module Raw : sig ... end

Raw operations for values of type TBitVector; be careful when using these, so as to provide properly typed values. Prefer using BitVec directly when possible, as it may also provide more reductions.

val contains_bvs : t -> (bool * int) option

Some (signed, size) if the given value contains somewhere some bitvector operations; None otherwise.

val contains_bvs_2 : t -> t -> (bool * int) option

val of_float : RoundingMode.t -> bool -> int -> t -> t

val of_int : bool -> int -> t -> t

of_int signed size v converts the value v of type TInt into an equivalent value of type TBitVector size, assuming the representation is signed.

val to_int : bool -> t -> t

val to_float : RoundingMode.t -> bool -> t -> t

val and_ : size:int -> signed:bool -> t -> t -> t

val or_ : size:int -> signed:bool -> t -> t -> t

val xor : size:int -> signed:bool -> t -> t -> t

val not : size:int -> signed:bool -> t -> t

val wrap_binop : (t -> t -> t) -> size:int -> signed:bool -> t -> t -> t

val shl : size:int -> signed:bool -> t -> t -> t

val ashr : size:int -> signed:bool -> t -> t -> t

val lshr : size:int -> signed:bool -> t -> t -> t

val wrap_plus : size:int -> signed:bool -> t -> t -> t

val wrap_minus : size:int -> signed:bool -> t -> t -> t

val wrap_times : size:int -> signed:bool -> t -> t -> t

val apply_binop_signed : (bool -> t -> t -> 'a) -> size:int -> signed:bool -> t -> t -> 'a

val plus_overflows : size:int -> signed:bool -> t -> t -> t

val times_overflows : size:int -> signed:bool -> t -> t -> t

val lt_as_bv : size:int -> signed:bool -> t -> t -> t

val leq_as_bv : size:int -> signed:bool -> t -> t -> t

val minus_overflows : size:int -> signed:bool -> t -> t -> t