Module BitVec.Raw

Raw operations for values of type TBitVector; be careful when using these, so as to provide properly typed values. Prefer using BitVec directly when possible, as it may also provide more reductions.

val mk : int -> Z.t -> t

val mk_masked : int -> Z.t -> t

val mk_like : t_node Hc.hash_consed -> Z.t -> t

val bv_to_z : bool -> int -> Z.t -> Z.t

bv_to_z signed bits z parses a BitVector z, for a given bitwidth bits, with signed, into an integer.

val max_for : bool -> int -> Z.t

max_for signed n is the inclusive maximum for a bitvector of size n when it is signed

val min_for : bool -> int -> Z.t

min_for signed n is the inclusive minimum for a bitvector of size n when it is signed

val is_2pow : Z.t -> bool

is_2pow z is true if z is a power of 2 greater than one. The > 1 check is done to ensure the mask for that number (ie. z - 1) is nonzero.

val is_left_mask : int -> Z.t -> bool

is_left_mask bits z is true if z is of the form 1+0+ and its length is bits

val is_right_mask : Z.t -> bool

is_right_mask z is true if z is of the form 0*1+

val right_mask_size : Z.t -> int

right_mask_size z is, for a z of the form 0*1{n}, n. If z is not of the form 0*1{n}, the result is undefined, so use is_right_mask before.

val covers_bitwidth : int -> Z.t -> bool

covers_bitwidth bits z is true if z is of the form 1+ and covers the whole bitwidth of size bits.

val plus : t -> t -> t

val minus : t -> t -> t

val times : t -> t -> t

val div : bool -> t -> t -> t

val rem : bool -> t -> t -> t

val mod_ : t -> t -> t

val not : t -> t

val and_ : t -> t -> t

val or_ : t -> t -> t

val xor : t -> t -> t

val extract : int -> int -> t -> t

val extend : bool -> int -> t -> t

val concat : t -> t -> t

concat v1 v2 for v1 of size n and v2 of size m is a bitvector of size n + m where the first m bits are v2 and the following n are v1

val shl : t -> t -> t

val lshr : t -> t -> t

val ashr : t -> t -> t

val lt : bool -> t -> t -> t

val leq : bool -> t -> t -> t

val gt : bool -> t -> t -> t

val geq : bool -> t -> t -> t

val plus_overflows : bool -> t -> t -> t

val neg_overflows : t -> t

val times_overflows : bool -> t -> t -> t