Module Soteria.Data

Symbolic abstractions over common data structures.

module S_bool : sig ... end

Symbolic abstraction over booleans.

module S_int : sig ... end

Symbolic abstraction over integers.

module S_range : sig ... end

Symbolic abstractions over ranges.

module S_map : sig ... end

A symbolic map abstraction. The main function of the symbolic map module type is find_opt, which allows symbolically looking up a key in the map (possibly branching), and returns the value at that key if any, along with the key to be used for future concrete operations on the map.