Module Range_tree.With_monad3

Parameters

module M : sig ... end

Signature

val map_leaves : ('a -> ('a, 'b, 'c) M.t) -> ('a, 'sint) t -> (('a, 'sint) t, 'b, 'c) M.t

map_leaves f t receives a mapping function f operating within the monad M, as well as a tree, and applys f to all leaves of the tree (not the inner nodes).