Module S_int.S

Parameters

module Symex : Symex.Base

module S_bool : S_bool.S(Symex).S

Signature

module type S = sig ... end

Symbolic integers, with basic operations: addition, subtraction, and comparison.

module type Bounded_S = sig ... end

Symbolic integers that are bounded: there is some invariant about them being in a range which must be upheld. This can for instance be used to represent machine integers (which have a min and max value) with numerals.