Module S_map.Concrete

Creates a "symbolic" map that relies purely on syntactic equality. Only sound to use if the keys of the map are invariant under interpretations of the symbolic variables. See Mk_concrete_key for a way to lift concrete keys into the symbolic realm.

Parameters

module Symex : Symex.Base

module Key : Key(Symex).S

Signature

type 'a t

val empty : 'a t

val is_empty : 'a t -> bool

val syntactic_bindings : 'a t -> (Key.t * 'a) Stdlib.Seq.t

val syntactic_mem : Key.t -> 'a t -> bool

val syntactic_add : Key.t -> 'a -> 'a t -> 'a t

val syntactic_add_opt : Key.t -> 'a option -> 'a t -> 'a t

val pp' : ?key:(Stdlib.Format.formatter -> Key.t -> unit) -> ?ignore:((Key.t * 'a) -> bool) -> (Stdlib.Format.formatter -> 'a -> unit) -> Stdlib.Format.formatter -> 'a t -> unit

val pp : (Stdlib.Format.formatter -> 'a -> unit) -> Stdlib.Format.formatter -> 'a t -> unit

val show : ('a -> string) -> 'a t -> string

val find_opt : Key.t -> 'a t -> (Key.t * 'a option) Symex.t

val fold : ('acc -> (Key.t * 'a) -> 'acc Symex.t) -> 'acc -> 'a t option -> 'acc Symex.t