Module Soteria.Logs

Logging facilities for symbolic execution.

This module (and particularly L) contains logging facilities for users of Soteria. The library knows how to produce text logs to stderr or structured logs in HTML format.

type ('a, 'b) msgf = (('a, Stdlib.Format.formatter, unit, 'b) Stdlib.format4 -> 'a) -> 'b

module Level : sig ... end

module Config : sig ... end

module L : sig ... end

module Profile : sig ... end

Information on the current printing profile, to know what capabilities the current output channel has.

module Color : sig ... end

Colors and styles that can be used in logging.

module Printers : sig ... end

Utility printers, that are aware of the current logging configuration.

module Import : sig ... end

Helper module to be opened when using logging.