Module Solvers.Smt_utils

SMT-LIB utilities for operations missing from Simple_smt.

Contains utilities for SMT-LIB's FloatingPoint theory and FixedSizeBitVecs theory, as well as some solver commands.

Rounding Modes

module RoundingMode : sig ... end

FloatingPoint

val float_shape : int -> int list

float_shape n is the shape of a IEEE float of a given size in bits. Returns a two element list [exp, mant], where exp is the number of exponent bits, and mant is the number of mantissa/significand bits. mant includes the hidden bit which is always 1, as per SMT-lib's expectations. Always holds that n = mant + exp. Only implemented for n = 16, 32, 64, 128.

val t_f16 : Simple_smt.sexp

SMT-LIB Float16 type.

val t_f32 : Simple_smt.sexp

SMT-LIB Float32 type.

val t_f64 : Simple_smt.sexp

SMT-LIB Float64 type.

val t_f128 : Simple_smt.sexp

SMT-LIB Float128 type.

val f32_k : float -> Simple_smt.sexp

f32_k f creates a Float32 constant from an OCaml float f.

Directly encodes the IEEE 754 binary32 representation. This may incur some loss of precision.

val f64_k : float -> Simple_smt.sexp

f64_k f creates a Float64 constant from an OCaml float f.

Directly encodes the IEEE 754 binary64 representation, so no precision is lost.

val f128_k : float -> Simple_smt.sexp

f128_k f creates a Float128 constant from OCaml float f.

The value is first converted into a Float64, and we then use the to_fp SMT-LIB function to convert it to a Float128, using RoundingMode.default. Necessarily implies a loss of precision.

val f16_k : float -> Simple_smt.sexp

f16_k f creates a Float16 constant from OCaml float f.

The value is first converted into a Float32, and we then use the to_fp SMT-LIB function to convert it to a Float16, using RoundingMode.default. Necessarily implies a loss of precision.

val fp_abs : Simple_smt.sexp -> Simple_smt.sexp

fp_abs f returns the absolute value of f.

val fp_eq : Simple_smt.sexp -> Simple_smt.sexp -> Simple_smt.sexp

fp_eq f1 f2 returns true if f1 equals f2 (IEEE 754 equality).

val fp_leq : Simple_smt.sexp -> Simple_smt.sexp -> Simple_smt.sexp

fp_leq f1 f2 returns true if f1 <= f2.

val fp_lt : Simple_smt.sexp -> Simple_smt.sexp -> Simple_smt.sexp

fp_lt f1 f2 returns true if f1 < f2.

val fp_add : Simple_smt.sexp -> Simple_smt.sexp -> Simple_smt.sexp

fp_add f1 f2 returns f1 + f2 using the default rounding mode (see RoundingMode.default).

val fp_sub : Simple_smt.sexp -> Simple_smt.sexp -> Simple_smt.sexp

fp_sub f1 f2 returns f1 - f2 using the default rounding mode (see RoundingMode.default).

val fp_mul : Simple_smt.sexp -> Simple_smt.sexp -> Simple_smt.sexp

fp_mul f1 f2 returns f1 * f2 using the default rounding mode (see RoundingMode.default).

val fp_div : Simple_smt.sexp -> Simple_smt.sexp -> Simple_smt.sexp

fp_div f1 f2 returns f1 / f2 using the default rounding mode (see RoundingMode.default).

val fp_rem : Simple_smt.sexp -> Simple_smt.sexp -> Simple_smt.sexp

fp_rem f1 f2 returns f1 % f2 (no rounding mode is involved here).

val fp_is : Stdlib.fpclass -> Simple_smt.sexp -> Simple_smt.sexp

fp_is fc f tests if f belongs to floating-point class fc (which is of OCaml's builtin fpclass type).

val fp_round : RoundingMode.t -> Simple_smt.sexp -> Simple_smt.sexp

fp_round rm f rounds f to an integer using rounding mode rm.

FloatingPoint - BitVec conversions

val float_of_bv : int -> Simple_smt.sexp -> Simple_smt.sexp

float_of_bv size bv interprets bitvector bv as a float of size bits. size must be one of 16, 32, 64 or 128.

This is a bitwise reinterpretation, not a numeric conversion.

val float_of_ubv : RoundingMode.t -> int -> Simple_smt.sexp -> Simple_smt.sexp

float_of_ubv rm size bv converts unsigned bitvector bv to a float, with rounding mode rm. size must be one of 16, 32, 64 or 128.

val float_of_sbv : RoundingMode.t -> int -> Simple_smt.sexp -> Simple_smt.sexp

float_of_ubv rm size bv converts signed bitvector bv to a float, with rounding mode rm. size must be one of 16, 32, 64 or 128.

val ubv_of_float : RoundingMode.t -> int -> Simple_smt.sexp -> Simple_smt.sexp

ubv_of_float rm size f converts float f to an unsigned bitvector of size bits, with rounding mode rm. This is a numeric conversion, not a bitwise reinterpretation. If f is out of range for the target bitvector, or NaN or inf, the result is undefined.

val sbv_of_float : RoundingMode.t -> int -> Simple_smt.sexp -> Simple_smt.sexp

ubv_of_float rm size f converts float f to a signed bitvector of size bits, with rounding mode rm. This is a numeric conversion, not a bitwise reinterpretation. If f is out of range for the target bitvector, or NaN or inf, the result is undefined.

Int - BitVec conversions

val int_of_bv : bool -> Simple_smt.sexp -> Simple_smt.sexp

int_of_bv signed bv converts bitvector bv to an integer. If signed is true, bv is interpreted as a signed bitvector; otherwise, it is interpreted as unsigned.

val bv_of_int : int -> Simple_smt.sexp -> Simple_smt.sexp

bv_of_int size n converts integer n to a bitvector of size bits.

BitVec

val bv_nego : Simple_smt.sexp -> Simple_smt.sexp

bv_nego x is true if negating x would overflow (signed).

val bv_uaddo : Simple_smt.sexp -> Simple_smt.sexp -> Simple_smt.sexp

bv_uaddo l r returns true if l + r would overflow (unsigned).

val bv_saddo : Simple_smt.sexp -> Simple_smt.sexp -> Simple_smt.sexp

bv_saddo l r returns true if l + r would overflow (signed).

val bv_usubo : Simple_smt.sexp -> Simple_smt.sexp -> Simple_smt.sexp

bv_usubo l r returns true if l - r would underflow (unsigned).

val bv_ssubo : Simple_smt.sexp -> Simple_smt.sexp -> Simple_smt.sexp

bv_ssubo l r returns true if l - r would underflow or overflow (signed).

val bv_umulo : Simple_smt.sexp -> Simple_smt.sexp -> Simple_smt.sexp

bv_umulo l r returns true if l * r would overflow (unsigned).

val bv_smulo : Simple_smt.sexp -> Simple_smt.sexp -> Simple_smt.sexp

bv_umulo l r returns true if l * r would overflow (signed).

Commands

val reset : Simple_smt.sexp

SMT-LIB (reset) command to reset the solver state.