Module type Value.S

The type of values the solver operators on.

This is different from Symex.Value.S, which represents symbolic values in the context of symbolic execution. Here we are simply interested in how some arbitrary typed values can be encoded into SMT-LIB.

type t

type ty

val init_commands : Simple_smt.sexp list

A list of commands that need to be aknowledged on start up of the solver. These are also sent everytime Solver_interface.S.reset is called.

val sort_of_ty : ty -> Simple_smt.sexp

Encode a type into a SMT-LIB sort

val encode_value : t -> Simple_smt.sexp

Encode a value into a SMT-LIB value