Module Z3.Make

Create a Z3 solver module from a value type. It can be configured, see Config.

Parameters

module Value : Value.S

Signature

type t

The solver state type.

type value = Value.t

The type of symbolic expressions.

type ty = Value.ty

The type of value types (for variable declarations).

val init : unit -> t

init () creates a new solver instance.

val add_constraint : t -> value -> unit

add_constraint solver v adds constraint v to the solver.

val check_sat : t -> Soteria.Symex.Solver_result.t

check_sat solver checks satisfiability of current constraints. This may return Symex.Solver_result.t.Unknown if the solver cannot determine satisfiability or if it times out.

val declare_var : t -> Soteria.Symex.Var.t -> ty -> unit

declare_var solver var ty declares a variable with type ty.

val push : t -> int -> unit

push solver n pushes n levels onto the assertion stack.

Constraints added after push can be removed by pop.

val pop : t -> int -> unit

pop solver n pops n levels from the assertion stack.

Removes all constraints added since the corresponding push.

val reset : t -> unit

reset solver clears all constraints and declarations.

Returns the solver to its initial state.