Module Soteria_std.Bimap

A bidirectional map implementation.

module Make (KeyL : Stdlib.Map.OrderedType) (KeyR : Stdlib.Map.OrderedType) : sig ... end

This functor builds a data structure that maintains a one-to-one correspondence between keys of type KeyL (left) and KeyR (right). It allows efficient lookup in both directions by internally maintaining two synchronized maps.

module MakePp (KeyL : Soteria_std.Ordered_type.S) (KeyR : Soteria_std.Ordered_type.S) : sig ... end

Equivalent to Make, but with pretty-printing.