Module Bimap.Make

This functor builds a data structure that maintains a one-to-one correspondence between keys of type KeyL (left) and KeyR (right). It allows efficient lookup in both directions by internally maintaining two synchronized maps.

Parameters

module KeyL : Stdlib.Map.OrderedType

module KeyR : Stdlib.Map.OrderedType

Signature

module M : sig ... end

module M_rev : sig ... end

type t = KeyR.t M.t * KeyL.t M_rev.t

val empty : 'a M.t * 'b M_rev.t

The empty map.

val is_empty : ('a M.t * 'b) -> bool

Test whether the map is empty.

val cardinal : ('a M.t * 'b) -> int

The number of bindings in the map.

val bindings : ('a M.t * 'b) -> (M.key * 'a) list

The list of all bindings from left to right keys.

val bindings_rev : ('a * 'b M_rev.t) -> (M_rev.key * 'b) list

The list of all bindings from right to left keys.

val add : M.key -> M_rev.key -> (M_rev.key M.t * M.key M_rev.t) -> M_rev.key M.t * M.key M_rev.t

Add a binding between a left key and a right key to the map.

val remove_l : M.key -> (M_rev.key M.t * 'a M_rev.t) -> M_rev.key M.t * 'a M_rev.t

Remove the binding associated with the given left key from the map.

val remove_r : M_rev.key -> ('a M.t * M.key M_rev.t) -> 'a M.t * M.key M_rev.t

Remove the binding associated with the given right key from the map.

val mem_l : M.key -> ('a M.t * 'b) -> bool

Whether the map contains a binding for the given left key.

val mem_r : M_rev.key -> ('a * 'b M_rev.t) -> bool

Whether the map contains a binding for the given right key.

val find_l : M.key -> ('a M.t * 'b) -> 'a option

Find the right key associated with the given left key.

val find_r : M_rev.key -> ('a * 'b M_rev.t) -> 'b option

Find the left key associated with the given right key.