Module Soteria_std.Cmdliner_helpers

Cmdliner argument converters for file and directory paths.

val file_as_absolute : unit -> string Cmdliner.Arg.Conv.t

Converter for file paths. Normalizes the path to an absolute path and checks if the file exists.

val dir_as_absolute : unit -> string Cmdliner.Arg.Conv.t

Converter for directory paths. Normalizes the path to an absolute path and checks if the directory exists.

val file_or_dir_as_absolute : [ `Dir of string | `File of string ] Cmdliner.Arg.Conv.t

Converter for file or directory paths. Normalizes the path to an absolute path and checks if it exists, returning either a file or a directory.

val regex : Str.regexp Cmdliner.Arg.Conv.t

module Sections : sig ... end

Common section names for command-line arguments. These can be used to group related options together in the help output.