Module Soteria_std.Effects

module type Bookkeeping = sig ... end

Module type for an effect used for bookkeeping, which keeps tracks of some information that can then be dumped "somewhere" (typically a file, configured externally), or ignored.