Module type Effects.Bookkeeping

Module type for an effect used for bookkeeping, which keeps tracks of some information that can then be dumped "somewhere" (typically a file, configured externally), or ignored.

type t

The type of the information tracked by the effect.

type arg

The type of the argument to the effect, if it needs to be tracked.

val with_ : arg -> (unit -> 'a) -> 'a * t

with_ arg f runs function f and handles effects raised by the functions of this module, and returns a pair of the result of f and the data tracked by the effect. This is to be used when the user wants to directly manage the data tracked by the effect.

val with_dumped : arg -> (unit -> 'a) -> 'a

with_dumped arg f runs function f and handles effects raised by the functions of this module, and dumps the resulting data to the file specified by the current (externally-defined) configuration if it is set, or ignores them otherwise (if configured to do so).

val with_ignored : unit -> (unit -> 'a) -> 'a

with_ignored () f runs function f and handles effects raised by the functions of this module, but ignores their effect. This is to be used when the user does not wish to pay the performance cost of tracking the effect.