Module Graph.Make_in_place

Parameters

module Node : Hashset.PrintableHashedType

Signature

type node = Node.t

module Node_set : sig ... end

module Hashtbl : sig ... end

type t = Node_set.t Hashtbl.t

val with_node_capacity : int -> 'a Hashtbl.t

Create a new graph with the given initial capacity for nodes.

val node_count : 'a Hashtbl.t -> int

Return the number of nodes in the graph.

val pp : Node_set.t Hashtbl.t Fmt.t

Pretty-printer for the graph.

val add_edge : Node_set.t Hashtbl.t -> Hashtbl.key -> Node_set.elt -> unit

Add a directed edge from one node to another.

val add_double_edge : Node_set.t Hashtbl.t -> Hashtbl.key -> Node_set.elt -> unit

Add a bidirectional edge between two nodes.

val set_edges_from : 'a Hashtbl.t -> Hashtbl.key -> 'a -> unit

Set the outgoing edges for a specific node.

val reachable_from : Node_set.t Hashtbl.t -> Node_set.t -> Node_set.t

Compute the set of all nodes reachable from a given set of start nodes.

val weak_topological_order : t -> node list

A topological order where SCCs are not necessarily well-ordered