Module Soteria_std.Hashset

A hash set implementation.

type 'a t

val with_capacity : int -> 'a t

Create a new empty set with the given initial capacity.

val singleton : 'a -> 'a t

Create a set containing a single element.

val add : 'a t -> 'a -> unit

Add an element to the set.

val add_iter : 'a t -> 'a Iter.t -> unit

Add elements from an iterator to the set.

val mem : 'a t -> 'a -> bool

Test whether an element is a member of the set.

val remove : 'a t -> 'a -> unit

Remove an element from the set.

val iter : ('a -> unit) -> 'a t -> unit

Iterate over all elements in the set.

val to_seq : 'a t -> 'a Soteria_std.Seq.t

Return a sequence of all elements in the set.

val of_seq : 'a Soteria_std.Seq.t -> 'a t

Create a set from a sequence of elements.

val cardinal : 'a t -> int

Return the number of elements in the set.

val copy : 'a t -> 'a t

Return a copy of the set.

val subseteq : 'a t -> 'a t -> bool

Test whether the first set is a subset of the second set.

val equal : 'a t -> 'a t -> bool

Test whether two sets are equal.

val pp : (Stdlib.Format.formatter -> 'a -> unit) -> Stdlib.Format.formatter -> 'a t -> unit

Pretty-printer for the set.

module type S = sig ... end

module type PrintableHashedType = sig ... end

module Make (Elt : PrintableHashedType) : S with type elt = Elt.t

Functor to create a hash set module for a given element type.

module Hstring : S with type elt = string

Hash set implementation for strings.

module Hint : S with type elt = int

Hash set implementation for integers.