Module Soteria_std.Iter

include module type of struct include Iter end

type !'a t = ('a -> unit) -> unit

type !'a iter = 'a t

type !'a equal = 'a -> 'a -> bool

type !'a hash = 'a -> int

val from_iter : (('a -> unit) -> unit) -> 'a t

val from_labelled_iter : (f:('a -> unit) -> unit) -> 'a t

val from_fun : (unit -> 'a option) -> 'a t

val empty : 'a t

val singleton : 'a -> 'a t

val doubleton : 'a -> 'a -> 'a t

val init : (int -> 'a) -> 'a t

val cons : 'a -> 'a t -> 'a t

val snoc : 'a t -> 'a -> 'a t

val return : 'a -> 'a t

val pure : 'a -> 'a t

val repeat : 'a -> 'a t

val iterate : ('a -> 'a) -> 'a -> 'a t

val forever : (unit -> 'b) -> 'b t

val cycle : 'a t -> 'a t

val unfoldr : ('b -> ('a * 'b) option) -> 'b -> 'a t

val scan : ('b -> 'a -> 'b) -> 'b -> 'a t -> 'b t

val iter : ('a -> unit) -> 'a t -> unit

val iteri : (int -> 'a -> unit) -> 'a t -> unit

val for_each : 'a t -> ('a -> unit) -> unit

val for_eachi : 'a t -> (int -> 'a -> unit) -> unit

val fold : ('a -> 'b -> 'a) -> 'a -> 'b t -> 'a

val foldi : ('a -> int -> 'b -> 'a) -> 'a -> 'b t -> 'a

val fold_map : ('acc -> 'a -> 'acc * 'b) -> 'acc -> 'a t -> 'b t

val fold_filter_map : ('acc -> 'a -> 'acc * 'b option) -> 'acc -> 'a t -> 'b t

val map : ('a -> 'b) -> 'a t -> 'b t

val mapi : (int -> 'a -> 'b) -> 'a t -> 'b t

val map_by_2 : ('a -> 'a -> 'a) -> 'a t -> 'a t

val for_all : ('a -> bool) -> 'a t -> bool

val exists : ('a -> bool) -> 'a t -> bool

val mem : ?eq:('a -> 'a -> bool) -> 'a -> 'a t -> bool

val find : ('a -> 'b option) -> 'a t -> 'b option

val find_map : ('a -> 'b option) -> 'a t -> 'b option

val findi : (int -> 'a -> 'b option) -> 'a t -> 'b option

val find_mapi : (int -> 'a -> 'b option) -> 'a t -> 'b option

val find_pred : ('a -> bool) -> 'a t -> 'a option

val find_pred_exn : ('a -> bool) -> 'a t -> 'a

val length : 'a t -> int

val is_empty : 'a t -> bool

val filter : ('a -> bool) -> 'a t -> 'a t

val append : 'a t -> 'a t -> 'a t

val append_l : 'a t list -> 'a t

val concat : 'a t t -> 'a t

val flatten : 'a t t -> 'a t

val flat_map : ('a -> 'b t) -> 'a t -> 'b t

val flat_map_l : ('a -> 'b list) -> 'a t -> 'b t

val seq_list : 'a t list -> 'a list t

val seq_list_map : ('a -> 'b t) -> 'a list -> 'b list t

val filter_map : ('a -> 'b option) -> 'a t -> 'b t

val filter_mapi : (int -> 'a -> 'b option) -> 'a t -> 'b t

val filter_count : ('a -> bool) -> 'a t -> int

val intersperse : 'a -> 'a t -> 'a t

val keep_some : 'a option t -> 'a t

val keep_ok : ('a, 'b) Stdlib.result t -> 'a t

val keep_error : ('a, 'e) Stdlib.result t -> 'e t

val persistent : 'a t -> 'a t

val persistent_lazy : 'a t -> 'a t

val sort : ?cmp:('a -> 'a -> int) -> 'a t -> 'a t

val sort_uniq : ?cmp:('a -> 'a -> int) -> 'a t -> 'a t

val sorted : ?cmp:('a -> 'a -> int) -> 'a t -> bool

val group_succ_by : ?eq:('a -> 'a -> bool) -> 'a t -> 'a list t

val group_by : ?hash:('a -> int) -> ?eq:('a -> 'a -> bool) -> 'a t -> 'a list t

val count : ?hash:('a -> int) -> ?eq:('a -> 'a -> bool) -> 'a t -> ('a * int) t

val uniq : ?eq:('a -> 'a -> bool) -> 'a t -> 'a t

val product : 'a t -> 'b t -> ('a * 'b) t

val diagonal_l : 'a list -> ('a * 'a) t

val diagonal : 'a t -> ('a * 'a) t

val join : join_row:('a -> 'b -> 'c option) -> 'a t -> 'b t -> 'c t

val join_by : ?eq:'key equal -> ?hash:'key hash -> ('a -> 'key) -> ('b -> 'key) -> merge:('key -> 'a -> 'b -> 'c option) -> 'a t -> 'b t -> 'c t

val join_all_by : ?eq:'key equal -> ?hash:'key hash -> ('a -> 'key) -> ('b -> 'key) -> merge:('key -> 'a list -> 'b list -> 'c option) -> 'a t -> 'b t -> 'c t

val group_join_by : ?eq:'a equal -> ?hash:'a hash -> ('b -> 'a) -> 'a t -> 'b t -> ('a * 'b list) t

val inter : ?eq:'a equal -> ?hash:'a hash -> 'a t -> 'a t -> 'a t

val union : ?eq:'a equal -> ?hash:'a hash -> 'a t -> 'a t -> 'a t

val diff : ?eq:'a equal -> ?hash:'a hash -> 'a t -> 'a t -> 'a t

val subset : ?eq:'a equal -> ?hash:'a hash -> 'a t -> 'a t -> bool

val max : ?lt:('a -> 'a -> bool) -> 'a t -> 'a option

val max_exn : ?lt:('a -> 'a -> bool) -> 'a t -> 'a

val min : ?lt:('a -> 'a -> bool) -> 'a t -> 'a option

val min_exn : ?lt:('a -> 'a -> bool) -> 'a t -> 'a

val sum : int t -> int

val sumf : float t -> float

val head : 'a t -> 'a option

val head_exn : 'a t -> 'a

val take : int -> 'a t -> 'a t

val take_while : ('a -> bool) -> 'a t -> 'a t

val map_while : ('a -> [ `Return of 'b | `Stop | `Yield of 'b ]) -> 'a t -> 'b t

val fold_while : ('a -> 'b -> 'a * [ `Continue | `Stop ]) -> 'a -> 'b t -> 'a

val drop : int -> 'a t -> 'a t

val drop_while : ('a -> bool) -> 'a t -> 'a t

val rev : 'a t -> 'a t

val zip_i : 'a t -> (int * 'a) t

val fold2 : ('c -> 'a -> 'b -> 'c) -> 'c -> ('a * 'b) t -> 'c

val iter2 : ('a -> 'b -> unit) -> ('a * 'b) t -> unit

val map2 : ('a -> 'b -> 'c) -> ('a * 'b) t -> 'c t

val map2_2 : ('a -> 'b -> 'c) -> ('a -> 'b -> 'd) -> ('a * 'b) t -> ('c * 'd) t

val to_list : 'a t -> 'a list

val to_rev_list : 'a t -> 'a list

val of_list : 'a list -> 'a t

val on_list : ('a t -> 'b t) -> 'a list -> 'b list

val pair_with_idx : 'a t -> (int * 'a) t

val to_opt : 'a t -> 'a option

val to_array : 'a t -> 'a array

val of_array : 'a array -> 'a t

val of_array_i : 'a array -> (int * 'a) t

val array_slice : 'a array -> int -> int -> 'a t

val of_opt : 'a option -> 'a t

val of_seq : 'a Stdlib.Seq.t -> 'a t

val to_seq_persistent : 'a t -> 'a Stdlib.Seq.t

val to_stack : 'a Stdlib.Stack.t -> 'a t -> unit

val of_stack : 'a Stdlib.Stack.t -> 'a t

val to_queue : 'a Stdlib.Queue.t -> 'a t -> unit

val of_queue : 'a Stdlib.Queue.t -> 'a t

val hashtbl_add : ('a, 'b) Stdlib.Hashtbl.t -> ('a * 'b) t -> unit

val hashtbl_replace : ('a, 'b) Stdlib.Hashtbl.t -> ('a * 'b) t -> unit

val to_hashtbl : ('a * 'b) t -> ('a, 'b) Stdlib.Hashtbl.t

val of_hashtbl : ('a, 'b) Stdlib.Hashtbl.t -> ('a * 'b) t

val hashtbl_keys : ('a, 'b) Stdlib.Hashtbl.t -> 'a t

val hashtbl_values : ('a, 'b) Stdlib.Hashtbl.t -> 'b t

val of_str : string -> char t

val to_str : char t -> string

val concat_str : string t -> string

exception OneShotSequence

val of_in_channel : Stdlib.in_channel -> char t

val to_buffer : char t -> Stdlib.Buffer.t -> unit

val int_range : start:int -> stop:int -> int t

val int_range_dec : start:int -> stop:int -> int t

val int_range_by : step:int -> int -> int -> int t

val bools : bool t

val of_set : (module Stdlib.Set.S with type elt = 'a and type t = 'b) -> 'b -> 'a t

val to_set : (module Stdlib.Set.S with type elt = 'a and type t = 'b) -> 'a t -> 'b

type !'a gen = unit -> 'a option

val of_gen : 'a gen -> 'a t

val of_gen_once : 'a gen -> 'a t

val to_gen : 'a t -> 'a gen

module Set = Iter.Set

module Map = Iter.Map

val random_int : int -> int t

val random_bool : bool t

val random_float : float -> float t

val random_array : 'a array -> 'a t

val random_list : 'a list -> 'a t

val shuffle : 'a t -> 'a t

val shuffle_buffer : int -> 'a t -> 'a t

val sample : int -> 'a t -> 'a array

module Infix = Iter.Infix

val (--) : int -> int -> int t

val (--^) : int -> int -> int t

val (>>=) : 'a t -> ('a -> 'b t) -> 'b t

val (>|=) : 'a t -> ('a -> 'b) -> 'b t

val (<*>) : ('a -> 'b) t -> 'a t -> 'b t

val (<+>) : 'a t -> 'a t -> 'a t

val (let+) : 'a t -> ('a -> 'b) -> 'b t

val (and+) : 'a t -> 'b t -> ('a * 'b) t

val (let*) : 'a t -> ('a -> 'b t) -> 'b t

val (and*) : 'a t -> 'b t -> ('a * 'b) t

val pp_seq : ?sep:string -> (Stdlib.Format.formatter -> 'a -> unit) -> Stdlib.Format.formatter -> 'a t -> unit

val pp_buf : ?sep:string -> (Stdlib.Buffer.t -> 'a -> unit) -> Stdlib.Buffer.t -> 'a t -> unit

val to_string : ?sep:string -> ('a -> string) -> 'a t -> string

module IO = Iter.IO

val of_list_combine : 'a list -> 'b list -> (('a * 'b) -> unit) -> unit

val combine_list : 'a t -> 'b list -> (('a * 'b) -> unit) -> unit

combine_list i l will combine the iterator i with the list l, returning a new iterator over the values of both i and l. If l is smaller than i, will raise a Invalid_argument at the end of the iterator.

val repeati : int -> 'a -> ('a -> unit) -> unit

val repeatz : Z.t -> 'a -> ('a -> unit) -> unit

val group_pairs_by : ?compare:('k -> 'k -> int) -> ((('k * 'a) -> unit) -> unit) -> (('k * 'a list) -> unit) -> unit

Groups an iterator over pairs, using the provided comparison function for sorting the keys. Eagerly traverses the provided iterator.