Module Soteria_std.List

Extensions to Stdlib.List with utility functions.

include module type of struct include Stdlib.List end

type !'a t = 'a list =

| []

| :: of 'a * 'a list

val length : 'a list -> int

val compare_lengths : 'a list -> 'b list -> int

val compare_length_with : 'a list -> int -> int

val is_empty : 'a list -> bool

val cons : 'a -> 'a list -> 'a list

val singleton : 'a -> 'a list

val hd : 'a list -> 'a

val tl : 'a list -> 'a list

val nth : 'a list -> int -> 'a

val nth_opt : 'a list -> int -> 'a option

val rev : 'a list -> 'a list

val init : int -> (int -> 'a) -> 'a list

val append : 'a list -> 'a list -> 'a list

val rev_append : 'a list -> 'a list -> 'a list

val concat : 'a list list -> 'a list

val flatten : 'a list list -> 'a list

val equal : ('a -> 'a -> bool) -> 'a list -> 'a list -> bool

val compare : ('a -> 'a -> int) -> 'a list -> 'a list -> int

val iter : ('a -> unit) -> 'a list -> unit

val iteri : (int -> 'a -> unit) -> 'a list -> unit

val map : ('a -> 'b) -> 'a list -> 'b list

val mapi : (int -> 'a -> 'b) -> 'a list -> 'b list

val rev_map : ('a -> 'b) -> 'a list -> 'b list

val filter_map : ('a -> 'b option) -> 'a list -> 'b list

val concat_map : ('a -> 'b list) -> 'a list -> 'b list

val fold_left_map : ('acc -> 'a -> 'acc * 'b) -> 'acc -> 'a list -> 'acc * 'b list

val fold_left : ('acc -> 'a -> 'acc) -> 'acc -> 'a list -> 'acc

val fold_right : ('a -> 'acc -> 'acc) -> 'a list -> 'acc -> 'acc

val iter2 : ('a -> 'b -> unit) -> 'a list -> 'b list -> unit

val map2 : ('a -> 'b -> 'c) -> 'a list -> 'b list -> 'c list

val rev_map2 : ('a -> 'b -> 'c) -> 'a list -> 'b list -> 'c list

val fold_left2 : ('acc -> 'a -> 'b -> 'acc) -> 'acc -> 'a list -> 'b list -> 'acc

val fold_right2 : ('a -> 'b -> 'acc -> 'acc) -> 'a list -> 'b list -> 'acc -> 'acc

val for_all : ('a -> bool) -> 'a list -> bool

val exists : ('a -> bool) -> 'a list -> bool

val for_all2 : ('a -> 'b -> bool) -> 'a list -> 'b list -> bool

val exists2 : ('a -> 'b -> bool) -> 'a list -> 'b list -> bool

val mem : 'a -> 'a list -> bool

val memq : 'a -> 'a list -> bool

val find : ('a -> bool) -> 'a list -> 'a

val find_opt : ('a -> bool) -> 'a list -> 'a option

val find_index : ('a -> bool) -> 'a list -> int option

val find_map : ('a -> 'b option) -> 'a list -> 'b option

val find_mapi : (int -> 'a -> 'b option) -> 'a list -> 'b option

val filter : ('a -> bool) -> 'a list -> 'a list

val find_all : ('a -> bool) -> 'a list -> 'a list

val filteri : (int -> 'a -> bool) -> 'a list -> 'a list

val take : int -> 'a list -> 'a list

val drop : int -> 'a list -> 'a list

val take_while : ('a -> bool) -> 'a list -> 'a list

val drop_while : ('a -> bool) -> 'a list -> 'a list

val partition : ('a -> bool) -> 'a list -> 'a list * 'a list

val partition_map : ('a -> ('b, 'c) Stdlib.Either.t) -> 'a list -> 'b list * 'c list

val assoc : 'a -> ('a * 'b) list -> 'b

val assoc_opt : 'a -> ('a * 'b) list -> 'b option

val assq : 'a -> ('a * 'b) list -> 'b

val assq_opt : 'a -> ('a * 'b) list -> 'b option

val mem_assoc : 'a -> ('a * 'b) list -> bool

val mem_assq : 'a -> ('a * 'b) list -> bool

val remove_assoc : 'a -> ('a * 'b) list -> ('a * 'b) list

val remove_assq : 'a -> ('a * 'b) list -> ('a * 'b) list

val split : ('a * 'b) list -> 'a list * 'b list

val combine : 'a list -> 'b list -> ('a * 'b) list

val sort : ('a -> 'a -> int) -> 'a list -> 'a list

val stable_sort : ('a -> 'a -> int) -> 'a list -> 'a list

val fast_sort : ('a -> 'a -> int) -> 'a list -> 'a list

val sort_uniq : ('a -> 'a -> int) -> 'a list -> 'a list

val merge : ('a -> 'a -> int) -> 'a list -> 'a list -> 'a list

val to_seq : 'a list -> 'a Stdlib.Seq.t

val of_seq : 'a Stdlib.Seq.t -> 'a list

val take_count : int -> 'a t -> 'a t * int

Take a prefix of a list of at most n elements, or the entire list if the list is shorter than n.

returns
(prefix, count) where prefix is the prefix of the list and count is the number of elements taken.

raises Invalid_argument
if n is negative.

val combine_opt : 'a list -> 'b list -> ('a * 'b) list option

Combine two lists into a list of pairs.

val join_results : ('a, 'b) Stdlib.result list -> ('a list, 'b list) Stdlib.result

Aggregates results, returning all errors if any, otherwise all successful values.

val combine3 : 'a t -> 'b t -> 'c t -> ('a * 'b * 'c) t

Combine three lists into a list of triples.

val combine3_opt : 'a t -> 'b t -> 'c t -> ('a * 'b * 'c) t option

Combine three lists into a list of triples.

val combinei : 'a t -> 'b t -> (int * 'a * 'b) t

val map2i : (int -> 'a -> 'b -> 'c) -> 'a t -> 'b t -> 'c t

Map a function over two lists simultaneously, passing the current index as the first argument to the function. The function f receives arguments index element1 element2.

returns
the resulting list.

raises Invalid_argument
if the lists have different lengths.

val map_changed : ('a -> 'a) -> 'a list -> 'a list * bool

Map a function over a list and return both the resulting list and a boolean indicating whether any element was changed (using physical equality !=).

returns
(res, changed) where res is the resulting list and changed is a boolean indicating whether any element was changed.

val last : 'a t -> 'a

Return the last element of a list.

val last_opt : 'a t -> 'a option

Return the last element of a list, if any.

val concat_map2 : ('a -> 'b -> 'c t) -> 'a t -> 'b t -> 'c t

Apply a function to each pair of elements from two lists, concatenating all resulting lists.

returns
the resulting list.

raises Invalid_argument
if the lists have different lengths.

val prepend_concat_map2 : 'a t -> ('b -> 'c -> 'a t) -> 'b t -> 'c t -> 'a t

val sub : from:int -> len:int -> 'a list -> 'a list

sub ~from ~len l returns the sublist of l, from index from (inclusive), of length n.

val group_by : ?compare:('a -> 'a -> int) -> ('a * 'b) list -> ('a * 'b t) list

Group a list of pairs by their first component, using the provided comparison function to determine equality of keys. The resulting list contains pairs of keys and lists of corresponding values. The complexity is O(n log n), the list is traversed only once.