Module Soteria_std.PatriciaTree

Persistent maps and sets keyed by an integer-encodable type, implemented as big-endian patricia tries.

Structural sharing. Operations that do not change the container return a physically equal container, and operations that change it share untouched subtree with the input.

module type Key = sig ... end

Keys are identified by an integer, which determines their position in the trie.

module type Map = sig ... end

module type Set = sig ... end

module MakeMap (Key : Key) : Map with type key = Key.t

module MakeSet (Key : Key) : Set with type elt = Key.t

module MakeWeak (Key : Key) : Map with type key = Key.t

module MakeWeakSet (Key : Key) : Set with type elt = Key.t