Module Soteria_std.Reversible

module type Mutable = sig ... end

module Make_mutable (M : sig ... end) : sig ... end

module type In_place = sig ... end

module Mutable_to_in_place (M : Mutable) : sig ... end

module Make_in_place (M : sig ... end) : sig ... end

module type Immutable = sig ... end

module Effectful (M : sig ... end) : sig ... end

Effectful reversible computation. Note that this module is a functor, so that we can create any amount, we are not limited to a unique reversible effect.