Module Reversible.Effectful

Effectful reversible computation. Note that this module is a functor, so that we can create any amount, we are not limited to a unique reversible effect.

I.e. one can do

  module Rev1 = Effectful (struct
    type t = int
  end)

  module Rev2 = Effectful (struct
    type t = string
  end)

  let computation () =
    Rev1.save ();
    Rev2.save ();
    Rev1.backtrack ()

  let () = Rev1.run ~init:0 @@ fun () -> Rev2.run ~init:"x" @@ computation

Parameters

module M : sig ... end

Signature

type Stdlib.Effect.t +=

| Backtrack_n : int -> unit Stdlib.Effect.t

| Save : unit Stdlib.Effect.t

| Update : (M.t -> 'a * M.t) -> 'a Stdlib.Effect.t

val backtrack_n : int -> unit

val save : unit -> unit

val wrap : (M.t -> 'a * M.t) -> unit -> 'a

val wrap_read : (M.t -> 'a) -> unit -> 'a

val run : init:M.t -> (unit -> 'a) -> 'a