Module Reversible.Mutable_to_effectful

Converts a mutable reversible state to an effectful interface

Parameters

module M : Mutable

Signature

type t = M.t

Type of the reversible state. May be mutable!

val backtrack_n : int -> unit

Remove the last n checkpoints.

val save : unit -> unit

Save the current state as a new checkpoint.

val wrap : (t -> 'a) -> unit -> 'a

Apply function f to the current state, may modify the state in place if it is mutable.

val run : (unit -> 'a) -> 'a

Runs an effectful reversible computation f with a fresh mutable state.