Module type Reversible.Effectful

Interface for effectful reversible operations that operate on a state captured by an algebraic effect.

There is no init or reset here, initialisation and resetting is handled by run at the start and end of the computation.

type t

Type of the reversible state. May be mutable!

val backtrack_n : int -> unit

Remove the last n checkpoints.

val save : unit -> unit

Save the current state as a new checkpoint.

val wrap : (t -> 'a) -> unit -> 'a

Apply function f to the current state, may modify the state in place if it is mutable.

val run : (unit -> 'a) -> 'a

Runs an effectful reversible computation f with a fresh mutable state.