Module Soteria.Soteria_symex

module Approx : sig ... end

module Bi_abd : sig ... end

module Compo_res : sig ... end

module Excl : sig ... end

module Freeable : sig ... end

module Fuel_gauge : sig ... end

module Plist : sig ... end

module Pmap : sig ... end

module Pure_fun : sig ... end

State model component of pure functions. A piece of state represents a function `f()` that returns a value in a certain set. It is quite similar to the agreement algebra, except that there is no allocation, there necessarily exists a value, and if we don't know it, we instantiate it on read/consumption.

module Solver : sig ... end

module Solver_result : sig ... end

module Substs : sig ... end

module Symex : sig ... end

module Tree_block : sig ... end

module Value : sig ... end

module Var : sig ... end

module With_info : sig ... end