Module Soteria.Stats

Tracking of statistics across symbolic execution.

General definitions

type stat_entry =

| Int of int
(*
An integer statistic; joined by addition.
*)

| Float of float
(*
A float statistic; joined by addition.
*)

| StrSeq of string Stdlib.Dynarray.t
(*
A sequence of strings; joined by concatenation.
*)

| Map of stat_entry Soteria.Soteria_std.Hashtbl.Hstring.t
(*
A map from strings to stat entries; joined by merging entries recursively.
*)

| Yojson of Yojson.Safe.t
(*
A JSON value; joined by creating a JSON array of the two values if both are not arrays, or concatenating the two arrays if they are. Use this as a last resort when no other type fits.
*)

type t

val to_yojson : t -> Yojson.Safe.t

val of_yojson : Yojson.Safe.t -> t Ppx_deriving_yojson_runtime.error_or

type 'a with_stats = {

res : 'a;

stats : t;

}

exception Incompatible_entries

val with_empty_stats : 'a -> 'a with_stats

val map_with_stats : ('a -> 'b) -> 'a with_stats -> 'b with_stats

val merge : t -> t -> t

merge s1 s2 merges two statistics s1 and s2 into a single statistic by combining their entries. Does not modify s1 or s2.

val create : unit -> t

Creates a new, empty statistics collection.

Accessors

val as_int : stat_entry -> int

as_int entry returns the integer value of the statistic entry entry. Raises Incompatible_entries if entry is not an integer.

val as_float : stat_entry -> float

as_float entry returns the float value of the statistic entry entry. Raises Incompatible_entries if entry is not a float.

val as_strseq : stat_entry -> string Stdlib.Dynarray.t

as_strseq entry returns the string sequence value of the statistic entry entry. Raises Incompatible_entries if entry is not a string sequence.

val as_map : stat_entry -> stat_entry Soteria.Soteria_std.Hashtbl.Hstring.t

as_map entry returns the map value of the statistic entry entry. Raises Incompatible_entries if entry is not a map.

val as_yojson : stat_entry -> Yojson.Safe.t

as_yojson entry returns the JSON value of the statistic entry entry. Raises Incompatible_entries if entry is not a JSON value.

val get_entry : t -> string -> stat_entry option

get_entry stats name returns the statistic entry under the given name name in the statistics collection stats, or None if no such entry exists.

val get_int : ?default:int -> t -> string -> int

get_int ?default stats name returns the integer statistic under the given name name in the statistics collection stats. If no such entry exists, returns default if provided (defaults to 0). Raises Incompatible_entries if the entry under name is not an integer.

val get_float : ?default:float -> t -> string -> float

get_float ?default stats name returns the float statistic under the given name name in the statistics collection stats. If no such entry exists, returns default if provided (defaults to 0.0). Raises Incompatible_entries if the entry under name is not a float.

val get_strseq : t -> string -> string Stdlib.Dynarray.t

get_strseq stats name returns the string sequence statistic under the given name name in the statistics collection stats. If no such entry exists, returns an empty sequence. Raises Incompatible_entries if the entry under name is not a string sequence.

val get_map : t -> string -> stat_entry Soteria.Soteria_std.Hashtbl.Hstring.t

get_map stats name returns the map statistic under the given name name in the statistics collection stats. If no such entry exists, returns an empty map. Raises Incompatible_entries if the entry under name is not a map.

Printing, outputting

val default_printer : Stdlib.Format.formatter -> stat_entry -> unit

Default printer for statistics entries.

val pp_iter_list : (('a -> unit) -> 'b -> unit) -> (Stdlib.Format.formatter -> 'a -> unit) -> Stdlib.Format.formatter -> 'b -> unit

Utility printer to print a list given an iteration function. Exposed for convenience and consistency with the default stats printing.

val pp_iter_bindings_list : (('a -> 'b -> unit) -> 'c -> unit) -> (Stdlib.Format.formatter -> ('a * 'b) -> unit) -> Stdlib.Format.formatter -> 'c -> unit

Utility printer to print a list of bindings given an iteration function. Exposed for convenience and consistency with the default stats printing.

val register_printer : string -> ?name:string -> (t -> Stdlib.Format.formatter -> stat_entry -> unit) -> unit

register key ?name printer Registers a new printer for statistics entries of name name, possibly overriding previously set printers. If an entry with the given name is found during pretty-printing and no printer has been registered for that name, default_printer is used instead. printer also receives the full stats object, to enable more context-aware printing. A string name can be provided to be used as a header when printing the statistic; if none is provided, key is used instead.

val disable_printer : string -> unit

disable_printer key disables any printing for the statistic for key, meaning it will not be printed at all. Use register_printer to revert this.

val register_int_printer : string -> ?name:string -> (t -> Stdlib.Format.formatter -> int -> unit) -> unit

Convenience function to register a printer for Int statistics, fallsback to default_printer for other types.

val register_float_printer : string -> ?name:string -> (t -> Stdlib.Format.formatter -> float -> unit) -> unit

Convenience function to register a printer for Float statistics, fallsback to default_printer for other types.

val register_strseq_printer : string -> ?name:string -> (t -> Stdlib.Format.formatter -> string Stdlib.Dynarray.t -> unit) -> unit

Convenience function to register a printer for StrSeq statistics, fallsback to default_printer for other types.

val register_map_printer : string -> ?name:string -> (t -> Stdlib.Format.formatter -> stat_entry Soteria.Soteria_std.Hashtbl.Hstring.t -> unit) -> unit

Convenience function to register a printer for Map statistics, fallsback to default_printer for other types.

val pp : Stdlib.Format.formatter -> t -> unit

Pretty-prints statistics to the given formatter.

val output : t -> unit

Outputs the stats to a file, to stdout, or nowhere, according to the current configuration.

module As_ctx : sig ... end

Module for manipulating statistics as a context using algebraic effects (that are hidden). All calls to any function in this module should happen only inside a function wrapped with with_stats, ensuring that the statistics are properly passed around.

module Config : sig ... end

Configuration options for statistics output.