Module Sym_data.S_eq

Symbolic abstractions that support equality

module type S = sig ... end

Defines a type that supports symbolic semantic equality. For performance reasons, such a type most also support syntactic equality, such that syntactic equality (equal) implies semantic equality (sem_eq)

module Make_syntax (S : S) : sig ... end

module Of_concrete (Symex : Symex.S) (C : sig ... end) : S with module Symex = Symex and type t = C.t

Defines a type that supports symbolic semantic equality. For performance reasons, such a type most also support syntactic equality, such that syntactic equality (equal) implies semantic equality (sem_eq)