Module type S_eq.S

Defines a type that supports symbolic semantic equality. For performance reasons, such a type most also support syntactic equality, such that syntactic equality (equal) implies semantic equality (sem_eq)

module Symex : Symex.S

The symex world in which the type lives.

type t

val equal : t -> t -> bool

Syntactic equality, should be fast.

val sem_eq : t -> t -> Symex.Value.S_bool.t Symex.Value.t

Symbolic semantic equality

val distinct : t list -> Symex.Value.S_bool.t Symex.Value.t

Receives a list of syntactically different values and returns a symbolic boolean corresponding to the fact that these values are also semantically distinct.