Module Sym_data.S_map

Symbolic key-value maps.

module Key : sig ... end

module type S = sig ... end

Defines a type that is a symbolic map from key to value.

module Make (Symex : Symex.S) (Key : Key.S with module Symex = Symex) : S with type key = Key.t and module Symex = Symex

Defines a type that is a symbolic map from key to value.

module Make_failfast (Symex : Symex.S) (Key : Key.S with module Symex = Symex) : S with type key = Key.t and module Symex = Symex

Defines a type that is a symbolic map from key to value.

module Make_concrete (Symex : Symex.S) (K : Soteria_std.Ordered_type.S) : S with type key = K.t and module Symex = Symex

Defines a type that is a symbolic map from key to value.