Module type Key.S

include Stdlib.Map.OrderedType

type t

val compare : t -> t -> int

include S_elt.S with type t := t

include Soteria_std.Printable.S with type t := t

module Symex : Symex.S

val subst : (Soteria.Symex.Var.t -> Soteria.Symex.Var.t) -> t -> t

val iter_vars : t -> 'a Symex.Value.ty Soteria.Symex.Var.iter_vars

include S_eq.S with type t := t and module Symex := Symex

val equal : t -> t -> bool

Syntactic equality, should be fast.

val sem_eq : t -> t -> Symex.Value.S_bool.t Symex.Value.t

Symbolic semantic equality

val distinct : t list -> Symex.Value.S_bool.t Symex.Value.t

Receives a list of syntactically different values and returns a symbolic boolean corresponding to the fact that these values are also semantically distinct.

include Soteria_std.Printable.S with type t := t

val pp : Stdlib.Format.formatter -> t -> unit