Module S_map.Make_failfast

Defines a type that is a symbolic map from key to value.

Parameters

module Symex : Symex.S

module Key : Key.S with module Symex = Symex

Signature

module Symex = Symex

type key = Key.t

module Raw_map : Stdlib.Map.S with type key = key

type 'a t = 'a Raw_map.t

val find_opt : key -> 'a t -> (key * 'a option) Symex.t

find_opt key map returns a symbolic computation that yields

(key, None) if there is no key in the map that is semantically equal to key

(key', Some v) if there is a binding in the map for (key', v) and key' is semantically equal to key.

In both cases, the resulting key' can be used to insert or update the map directly without needing to search for it again.