Module type S_map.S

Defines a type that is a symbolic map from key to value.

module Symex : Symex.S

type key

module Raw_map : Stdlib.Map.S with type key = key

type 'a t = 'a Raw_map.t

val find_opt : key -> 'a t -> (key * 'a option) Symex.t

find_opt key map returns a symbolic computation that yields

(key, None) if there is no key in the map that is semantically equal to key

(key', Some v) if there is a binding in the map for (key', v) and key' is semantically equal to key.

In both cases, the resulting key' can be used to insert or update the map directly without needing to search for it again.