Module Sym_states.Pmap

module KeyS (Symex : Symex.Base) : sig ... end

module Mk_concrete_key (Symex : Symex.Base) (Key : Soteria_std.Ordered_type.S) : KeyS(Symex).S with type t = Key.t

module Build_from_find_opt_sym (Symex : Symex.Base) (Key : KeyS(Symex).S) (Find_opt_sym : sig ... end) : sig ... end

module Make (Symex : Symex.Base) (Key : KeyS(Symex).S) : sig ... end

module Direct_access (Symex : Symex.Base) (Key : KeyS(Symex).S) : sig ... end

module Concrete (Symex : Symex.Base) (Key : Soteria_std.Ordered_type.S) : sig ... end

Only sound to use if the keys of the map are invariant under interpretations of the symbolic variables