Module Symex.Stats

module Range : Stats.CodeRange

type t = Range.t Stats.stats

val to_yojson : t -> Yojson.Safe.t

val of_yojson : Yojson.Safe.t -> t Ppx_deriving_yojson_runtime.error_or

type nonrec 'a with_stats = ('a, Range.t) Stats.with_stats

val merge : t -> t -> t

Merges two stats records. Does not modify either.

val create : unit -> t

Empty record of statistics

val push_give_up_reason : loc:Range.t -> string -> t -> unit

Push a reason for giving up symex, with the location at which it was reached.

val push_missing_without_fix : string -> t -> unit

Push a reason for not being able to fix a Missing error

val did_give_up : t -> bool

Returns true if the stats record contains any give up reasons

val add_exec_time : float -> t -> unit

Adds the given execution time to the stats record.

val dump : t -> string -> unit

Dumps the stats record to a file in JSON format

val pp : Stdlib.Format.formatter -> t -> unit

Pretty-prints stats.

val output : t -> unit

Outputs the stats to a file, to stdout, or nowhere, according to the current configuration.

module As_ctx : sig ... end

Module for manipulating statistics as a context using algebraic effects (that are hidden). All calls to any function in this module should happen only inside a function wrapped with with_stats, ensuring that the statistics are properly passed around.