Module Expr.Subst

A susbtitution projects the syntactic to the semantic world.

type expr := t

type t

Type of substitutions

val pp : Stdlib.Format.formatter -> t -> unit

val empty : t

val apply : missing_var:(Symex.Var.t -> 'a ty -> 'a t) -> t -> expr -> 'a t * t

Applies a substitution to a syntactic representation to obtain a symbolic value. Should a variable binding be missing, fresh values are generated using the provided function.

val learn : t -> expr -> 'a t -> t option

learn θ e v takes the substitution θ, the expression e and the value v and tries to complete θ into a substitution θ' such that θ'(e) = v.

The only requirement from this function is that, if this function returns Some θ', then θ' covers all variables in e. However, all implementations of this function should be sound:

If this function returns a substitution that does not cover all variables in e, the client may crash (but not result to an unsoundness).

If this function returns a wrong substitution (where θ'(e) != v), this will not result in an unsoundness, but may prevent successful consumption.

If there are multiple substitutions that satisfy the requirement, this function can return any of them. This will lead to under-approximation when in UX mode, and to a probable consumption failure in UX mode.

If this function returns None when there exists a correct substitution, this will only lead to consumption failure. The function can always be extended to return a correct substitution.