Module Sym_states.Tree_block

module Split_tree : sig ... end

A Split_tree is a simplified representation of a tree, that has no offset. It however indicates, on Nodes, at what offset the split occurs, relative to that node's start.

type ('a, 'sint) tree = {

node : 'a node;

range : 'sint * 'sint;

children : (('a, 'sint) tree * ('a, 'sint) tree) option;

}

and 'a node =

| NotOwned of node_qty

| Owned of 'a

and node_qty =

| Partially

| Totally

val pp_tree : 'a 'sint. (Ppx_deriving_runtime.Format.formatter -> 'a -> Ppx_deriving_runtime.unit) -> (Ppx_deriving_runtime.Format.formatter -> 'sint -> Ppx_deriving_runtime.unit) -> Ppx_deriving_runtime.Format.formatter -> ('a, 'sint) tree -> Ppx_deriving_runtime.unit

val show_tree : 'a 'sint. (Ppx_deriving_runtime.Format.formatter -> 'a -> Ppx_deriving_runtime.unit) -> (Ppx_deriving_runtime.Format.formatter -> 'sint -> Ppx_deriving_runtime.unit) -> ('a, 'sint) tree -> Ppx_deriving_runtime.string

val pp_node : 'a. (Ppx_deriving_runtime.Format.formatter -> 'a -> Ppx_deriving_runtime.unit) -> Ppx_deriving_runtime.Format.formatter -> 'a node -> Ppx_deriving_runtime.unit

val show_node : 'a. (Ppx_deriving_runtime.Format.formatter -> 'a -> Ppx_deriving_runtime.unit) -> 'a node -> Ppx_deriving_runtime.string

val pp_node_qty : Ppx_deriving_runtime.Format.formatter -> node_qty -> Ppx_deriving_runtime.unit

val show_node_qty : node_qty -> Ppx_deriving_runtime.string

val make_tree : node:'a node -> range:('b * 'b) -> ?children:(('a, 'b) tree * ('a, 'b) tree) -> unit -> ('a, 'b) tree

module MemVal (Symex : Symex.Base) : sig ... end

The input module of Tree_block. A memory value t represents an owned part of the tree block, with the property that it can be split or merged as needed.

module Make (Symex : Symex.Base) (MemVal : MemVal(Symex).S) : sig ... end