Module Tree.Frame_range

Parameters

module M : sig ... end

Signature

val (let*) : ('a, 'b, 'c) M.t -> ('a -> ('d, 'b, 'c) M.t) -> ('d, 'b, 'c) M.t

val (let+) : ('a, 'b, 'c) M.t -> ('a -> 'd) -> ('d, 'b, 'c) M.t

val (let*^) : 'a Symex.t -> ('a -> ('b, 'c, 'd) M.t) -> ('b, 'c, 'd) M.t

val (let+^) : 'a Symex.t -> ('a -> 'b) -> ('b, 'c, 'd) M.t

val frame_range : t -> replace_node:(t -> (t, 'b, 'c) M.t) -> rebuild_parent: ((MemVal.t, sint) tree -> left:t -> right:t -> (MemVal.t, sint) tree Symex.t) -> Range.t -> (t * t, 'b, 'c) M.t

frame_range t ~replace_node ~rebuild_parent range Extracts from t the subtree that exactly spans range. The range must be non-empty. If t does not already cover range, it is first extended with NotOwned nodes so that it does.

Once the target subtree is found, replace_node is applied to it and must return the replacement subtree.

The path back to the root is then rebuilt by calling rebuild_parent with the (possibly modified) children at each step. Use of_children_s for rebuild_parent when the parent node needs to be recomputed from its children. Use with_children when only the tree structure changed (e.g. after a load) and no recomputation of the parent’s semantic content is required. In doubt, of_children_s is usually a safe bet.

frame_range returns a pair of the extracted subtree (before modification with replace_node) and the new root of the whole tree.