Module MemVal.S_bounded_int

include Soteria.Data.S_int.S(Symex)(S_bool).S

type +'a v := 'a Symex.Value.t

type t

val sem_eq : t v -> t v -> S_bool.t Symex.Value.t

Symbolic semantic equality

val of_z : Z.t -> t v

Takes an integer and creates an abstraction over it

val to_z : t v -> Z.t option

Takes a symbolic integer and returns Some z if this abstraction describes exactly a single integer, and None otherwise

val zero : unit -> t v

val one : unit -> t v

Arithmetic operations

val add : t v -> t v -> t v

val sub : t v -> t v -> t v

val lt : t v -> t v -> S_bool.t v

val leq : t v -> t v -> S_bool.t v

val is_in_bound : t Symex.Value.t -> S_bool.t Symex.Value.t

Checks whether the given symbolic integer is in the range described by the invariant.