Module MemVal.S_int

include Sym_data.S_elt.S with module Symex = Symex

type t

include Soteria_std.Printable.S with type t := t

module Symex = Symex

val subst : (Soteria.Symex.Var.t -> Soteria.Symex.Var.t) -> t -> t

val iter_vars : t -> 'a Symex.Value.ty Soteria.Symex.Var.iter_vars

include Sym_data.S_eq.S with module Symex := Symex and type t := t

val equal : t -> t -> bool

Syntactic equality, should be fast.

val sem_eq : t -> t -> Symex.Value.S_bool.t Symex.Value.t

Symbolic semantic equality

val distinct : t list -> Symex.Value.S_bool.t Symex.Value.t

Receives a list of syntactically different values and returns a symbolic boolean corresponding to the fact that these values are also semantically distinct.

include Soteria_std.Printable.S with type t := t

val pp : Stdlib.Format.formatter -> t -> unit

val of_z : Z.t -> t

Takes an integer and creates an abstraction over it

val to_z : t -> Z.t option

Takes a symbolic integer and returns Some z if this abstraction describes exactly a single integer, and None otherwise

val zero : unit -> t

val one : unit -> t

Arithmetic operations

type sbool_v := Symex.Value.S_bool.t Symex.Value.t

val add : t -> t -> t

val minus : t -> t -> t

val lt : t -> t -> sbool_v

val leq : t -> t -> sbool_v