Module Symex.Compo_res

This module defines a three-way result type for compositional symbolic execution. Unlike standard Result.t which has two cases (Ok/Error), this type adds a third case Missing for bi-abduction scenarios where anti-frame inference is needed, or more generally to represent incompletenesses in the engine.

When performing a function call, the current state may not contain all resources needed by the callee. Rather than immediately failing, we can infer what's missing (the "anti-frame") and continue analysis.

Three Cases:

Ok: Operation succeeded, resources matched

Error: Definite error found (e.g., null dereference, assertion failure)

Missing: Resources needed but not present; includes fixes. Missing contains a list of fixes, providing different ways to resolve the missing resources; for instance, when accessing a location in memory, two options are possible: either the location is allocated, or the location has been freed.

This module also provides a functor to lift this result type into any monadic context.

type ('ok, 'err, 'fix) t =

| Ok of 'ok

| Error of 'err

| Missing of 'fix list

val pp : ok:(Stdlib.Format.formatter -> 'a -> unit) -> err:(Stdlib.Format.formatter -> 'b -> unit) -> miss:'c Fmt.t -> Stdlib.Format.formatter -> ('a, 'b, 'c) t -> unit

val ok : 'a -> ('a, 'b, 'c) t

val error : 'a -> ('b, 'a, 'c) t

val miss : 'a list -> ('b, 'c, 'a) t

val is_ok : ('a, 'b, 'c) t -> bool

val is_error : ('a, 'b, 'c) t -> bool

val is_missing : ('a, 'b, 'c) t -> bool

val get_ok : ('a, 'b, 'c) t -> 'a

val get_error : ('a, 'b, 'c) t -> 'b

val get_missing : ('a, 'b, 'c) t -> 'c list

val only_oks : ('a, 'b, 'c) t list -> 'a list

val only_errors : ('a, 'b, 'c) t list -> 'b list

val only_missings : ('a, 'b, 'c) t list -> 'c list list

val bind : ('a, 'b, 'c) t -> ('a -> ('d, 'b, 'c) t) -> ('d, 'b, 'c) t

val map : ('a, 'b, 'c) t -> ('a -> 'd) -> ('d, 'b, 'c) t

val bind_error : ('a, 'b, 'c) t -> ('b -> ('a, 'd, 'c) t) -> ('a, 'd, 'c) t

val map_error : ('a, 'b, 'c) t -> ('b -> 'd) -> ('a, 'd, 'c) t

val map_missing : ('a, 'b, 'c) t -> ('c -> 'd) -> ('a, 'b, 'd) t

module Syntax : sig ... end

module T (M : Soteria.Soteria_std.Monad.Base) : sig ... end