Module Symex.Fuel_gauge

type exhaust =

| Exhausted

| Not_exhausted

module Fuel_value : sig ... end

type t = {

steps : Fuel_value.t;
(*
The number of steps that can be taken per branch.
*)

branching : Fuel_value.t;
(*
Number of times branching happens. Careful, the total number of branches is potentially exponential in the number of branchings
*)

}

val pp : Ppx_deriving_runtime.Format.formatter -> t -> Ppx_deriving_runtime.unit

val show : t -> Ppx_deriving_runtime.string

val infinite : t

val is_infinite : t -> bool

val consume_fuel_steps : int -> t -> exhaust * t

val consume_branching : int -> t -> exhaust * t

val take_branches : 'a Soteria.Soteria_std.List.t -> t -> 'a Soteria.Soteria_std.List.t * t

Receives a list and a fuel gauge and returns a list that contains at most as many elements as the fuel gauge allows. It also returns the updated fuel gauge.

module Cli : sig ... end