Module Symex.Solver

module type Mutable_incremental = sig ... end

This module represents a solver state, it is fully imperative!

module Mutable_to_effectful (M : Mutable_incremental) : sig ... end

Converts a mutable incremental solver into an effectful one. See the documentation of Soteria_std.Reversible.Mutable_to_effectful for more details.

module Mutable_to_pooled (M : Mutable_incremental) : sig ... end

Converts a mutable incremental solver into a pooled one. See the documentation of Soteria_std.Reversible.Mutable_to_pooled for more details.