Module Symex.Result

type nonrec ('ok, 'err, 'fix) t = ('ok, 'err, 'fix) Compo_res.t t

val ok : 'ok -> ('ok, 'err, 'fix) t

val error : 'err -> ('ok, 'err, 'fix) t

val miss : 'fix list -> ('ok, 'err, 'fix) t

val miss_no_fix : reason:string -> unit -> ('ok, 'err, 'fix) t

Missing without any fix. Will add to the statistics and log that information.

val bind : ('ok, 'err, 'fix) t -> ('ok -> ('a, 'err, 'fix) t) -> ('a, 'err, 'fix) t

val map : ('ok, 'err, 'fix) t -> ('ok -> 'a) -> ('a, 'err, 'fix) t

val bind_2 : ('ok, 'err, 'fix) t -> f:('ok -> ('a, 'b, 'fix) t) -> fe:('err -> ('a, 'b, 'fix) t) -> ('a, 'b, 'fix) t

val bind_error : ('ok, 'err, 'fix) t -> ('err -> ('ok, 'a, 'fix) t) -> ('ok, 'a, 'fix) t

val map_error : ('ok, 'err, 'fix) t -> ('err -> 'a) -> ('ok, 'a, 'fix) t

val map_missing : ('ok, 'err, 'fix) t -> ('fix -> 'a) -> ('ok, 'err, 'a) t

val fold_list : 'a list -> init:'b -> f:('b -> 'a -> ('b, 'c, 'd) t) -> ('b, 'c, 'd) t

val fold_iter : 'a Soteria_std.Iter.t -> init:'b -> f:('b -> 'a -> ('b, 'c, 'd) t) -> ('b, 'c, 'd) t

val fold_seq : 'a Soteria.Soteria_std.Seq.t -> init:'b -> f:('b -> 'a -> ('b, 'c, 'd) t) -> ('b, 'c, 'd) t

val iter_list : 'a list -> f:('a -> (unit, 'b, 'c) t) -> (unit, 'b, 'c) t

val iter_iter : 'a Soteria_std.Iter.t -> f:('a -> (unit, 'b, 'c) t) -> (unit, 'b, 'c) t

val map_list : 'a list -> f:('a -> ('b, 'c, 'd) t) -> ('b list, 'c, 'd) t

val run : ?fuel:Fuel_gauge.t -> ?fail_fast:bool -> mode:Approx.t -> ('ok, 'err, 'fix) t -> (('ok, 'err Or_gave_up.t, 'fix) Compo_res.t * Value.sbool Value.t list) list

Same as run, but receives a symbolic process that returns a Symex.Compo_res.t and maps the result to an Symex.Or_gave_up.t, potentially adding any path that gave up to the list.

val run_with_stats : ?fuel:Fuel_gauge.t -> ?fail_fast:bool -> mode:Approx.t -> ('ok, 'err, 'fix) t -> (('ok, 'err Or_gave_up.t, 'fix) Compo_res.t * Value.sbool Value.t list) list Stats.with_stats

val run_needs_stats : ?fuel:Fuel_gauge.t -> ?fail_fast:bool -> mode:Approx.t -> ('ok, 'err, 'fix) t -> (('ok, 'err Or_gave_up.t, 'fix) Compo_res.t * Value.sbool Value.t list) list