Module Symex.Var

type t

Variable names are integers under the hood

type 'ty iter_vars = ((t * 'ty) -> unit) -> unit

Iterator over variables and their types

val hash : t -> int

Hashes the variable name

val of_int : int -> t

Converts an integer into a variable. Identity function.

val to_int : t -> int

val to_string : t -> string

Sets up the variable name into an SMTlib-valid string name.

val of_string : string -> t

of_string [to_string t] is always t, but the function may raise an exception if the input was not obtained through to_string

val pp : Stdlib.Format.formatter -> t -> unit

val equal : t -> t -> bool

val compare : t -> t -> int

module Incr_counter_mut (_ : sig ... end) : sig ... end

module Hashset : Soteria.Soteria_std.Hashset.S with type elt = t

module Set : PatriciaTree.SET with type elt = t

module Hashtbl : Soteria.Soteria_std.Hashtbl.S with type key = t

module Map : PatriciaTree.MAP with type key = t