Module Var.Incr_counter_mut

Parameters

module _ : sig ... end

Signature

type var = t

type t

include Soteria.Soteria_std.Reversible.Mutable with type t := t

val init : unit -> t

Create a new reversible state initialized with the default value.

val backtrack_n : t -> int -> unit

Remove the last n checkpoints from state.

val save : t -> unit

Save the current state as a new checkpoint.

val reset : t -> unit

Clear all checkpoints and reset to the default value.

val get_next : t -> var