Module Terminal.Printers

val pp_unstable : name:string -> (Stdlib.Format.formatter -> 'a -> unit) -> Stdlib.Format.formatter -> 'a -> unit

Wraps a pretty-printer for unstable values, that should not be output into e.g. testing environments where diffs matter. If the config option hide_unstable is set, will replace

val pp_time : Stdlib.Format.formatter -> float -> unit

Pretty-prints a time quantity in seconds, displaying it in either seconds or milli-seconds for smaller values (< 0.05s). Wrapped in pp_unstable.

val pp_percent : Stdlib.Format.formatter -> (float * float) -> unit

For a given value (total, part), prints what percentage of total is represented by part. e.g. pp_percent stdout (5, 1) will print 20.00%

val pp_plural : sing:string -> plur:string -> Stdlib.Format.formatter -> int -> unit

Prints 1 sing if n = 1, n plur otherwise.