Module Tiny_values.Svalue

module Var = Soteria.Symex.Var

type ty =

| TBool

| TInt

val equal_ty : ty -> ty -> Ppx_deriving_runtime.bool

val pp_ty : Ppx_deriving_runtime.Format.formatter -> ty -> Ppx_deriving_runtime.unit

val show_ty : ty -> Ppx_deriving_runtime.string

val compare_ty : ty -> ty -> Ppx_deriving_runtime.int

val t_bool : ty

val t_int : ty

module Nop : sig ... end

module Unop : sig ... end

module Binop : sig ... end

val pp_hash_consed : ('a -> 'b -> 'c) -> 'a -> 'b Hc.hash_consed -> 'c

val equal_hash_consed : 'a -> 'b Hc.hash_consed -> 'c Hc.hash_consed -> bool

val compare_hash_consed : 'a -> 'b Hc.hash_consed -> 'c Hc.hash_consed -> int

type t_kind =

| Var of Var.t

| Bool of bool

| Int of Z.t

| Unop of Unop.t * t

| Binop of Binop.t * t * t

| Nop of Nop.t * t list

| Ite of t * t * t

and t_node = {

kind : t_kind;

ty : ty;

}

and t = t_node Hc.hash_consed

val pp_t_kind : Ppx_deriving_runtime.Format.formatter -> t_kind -> Ppx_deriving_runtime.unit

val show_t_kind : t_kind -> Ppx_deriving_runtime.string

val pp_t_node : Ppx_deriving_runtime.Format.formatter -> t_node -> Ppx_deriving_runtime.unit

val show_t_node : t_node -> Ppx_deriving_runtime.string

val show : t -> Ppx_deriving_runtime.string

val equal_t_kind : t_kind -> t_kind -> Ppx_deriving_runtime.bool

val equal_t_node : t_node -> t_node -> Ppx_deriving_runtime.bool

val compare_t_kind : t_kind -> t_kind -> Ppx_deriving_runtime.int

val compare_t_node : t_node -> t_node -> Ppx_deriving_runtime.int

val hash : 'a Hc.hash_consed -> int

val kind : t_node Hc.hash_consed -> t_kind

val iter_vars : t -> ((Var.t * ty) -> unit) -> unit

val pp_full : Ppx_deriving_runtime.Format.formatter -> t_node Hc.hash_consed -> Ppx_deriving_runtime.unit

val pp : t Fmt.t

val equal : 'a Hc.hash_consed -> 'b Hc.hash_consed -> bool

val compare : 'a Hc.hash_consed -> 'b Hc.hash_consed -> int

val sure_neq : t_node Hc.hash_consed -> t_node Hc.hash_consed -> bool

module Hcons : sig ... end

val (<|) : t_kind -> ty -> t

val mk_var : Var.t -> ty -> t

val mk_commut_binop : Binop.t -> t -> t -> t_kind

We put commutative binary operators in some sort of normal form where element with the smallest id is on the LHS, to increase cache hits.

val subst : (Var.t -> Var.t) -> t -> t

Booleans

val v_true : t

val v_false : t

val to_bool : 'a Hc.hash_consed -> bool option

val of_bool : bool -> t

val and_ : t -> t -> t

val or_ : t -> t -> t

val conj : t list -> t

val not : t -> t

val split_ands : t -> (t -> unit) -> unit

val distinct : t list -> t

val ite : t -> t -> t -> t

Integers

val int_z : Z.t -> t

val int : int -> t

val nonzero_z : Z.t -> t

val nonzero : int -> t

val zero : t

val one : t

val add : t -> t -> t

val sub : t -> t -> t

val mul : t -> t -> t

val div : t -> t -> t

val is_mod : t -> Z.t -> bool

val rem : t -> t -> t

val mod_ : t -> t -> t

val neg : t -> t

val lt : t -> t -> t

val leq : t -> t -> t

val geq : t -> t -> t

val gt : t -> t -> t

val sem_eq : t -> t -> t

val sem_eq_untyped : t -> t -> t

Infix operators

module Infix : sig ... end

module Syntax : sig ... end