Module type Symex.Base

module Value : Value.S

module MONAD : Soteria_std.Monad.Base

type 'a v := 'a Value.t

type 'a vt := 'a Value.ty

type sbool := Value.sbool

val assume : sbool v list -> unit MONAD.t

val vanish : unit -> 'a MONAD.t

val assert_ : sbool v -> bool MONAD.t

val assert_ox : sbool v -> bool MONAD.t

val nondet : 'a vt -> 'a v MONAD.t

val fresh_var : 'a vt -> Var.t MONAD.t

val branch_on : ?left_branch_name:string -> ?right_branch_name:string -> sbool v -> then_:(unit -> 'a MONAD.t) -> else_:(unit -> 'a MONAD.t) -> 'a MONAD.t

val branch_on_take_one : ?left_branch_name:string -> ?right_branch_name:string -> sbool v -> then_:(unit -> 'a MONAD.t) -> else_:(unit -> 'a MONAD.t) -> 'a MONAD.t

Branches on value, and takes at most one branch, starting with the then branch. This means that if the then_ branch is SAT, it is taken and the else_ branch is ignored, otherwise the else_ branch is taken. This is, of course, UX-sound, but not OX-sound.

val branches : (unit -> 'a MONAD.t) list -> 'a MONAD.t

Fuel

val consume_fuel_steps : int -> unit MONAD.t

val run : fuel:Fuel_gauge.t -> 'a MONAD.t -> ('a * sbool v list) list

run p actually performs symbolic execution and returns a list of obtained branches which capture the outcome together with a path condition that is a list of boolean symbolic values