Soteria documentation

Soteria is a comprehensive library for writing symbolic execution engines. The library is very parametric and allows for defining one's own notion of symbolic value or solver, enabling arbitrary language-specific optimisations. Our goal is for symbolic execution engines to be easy to write, easy to read, highly customisable and efficient.

Manual contents

The symbolic execution tutorial walks you through the basics of doing symbolic execution with Soteria.

Library

Soteria

Soteria.Symex

Soteria.Sym_states

Soteria.Logs Logging facilities for symbolic execution.

Soteria.Tiny_values

Soteria.BV_values

Soteria.Terminal

Soteria.Stats

Soteria.Soteria_std Standard library extensions and utilities for Soteria.