Module Bv_values.Bv_solver

module Z3_solver : Soteria.Symex.Solver.Mutable_incremental with module Value = Typed

This module represents a solver state, it is fully imperative!

module Z3_incremental_solver : Soteria.Symex.Solver.Mutable_incremental with module Value = Typed

This module represents a solver state, it is fully imperative!