Module Bv_values.Eval

val eval_binop : Svalue.Binop.t -> Svalue.t -> Svalue.t -> Svalue.t

val eval_unop : Svalue.Unop.t -> Svalue.t -> Svalue.t

val eval : ?force:bool -> ?eval_var:(Svalue.t -> Soteria.Symex.Var.t -> Svalue.ty -> Svalue.t) -> Svalue.t -> Svalue.t

Evaluates an expression; will call eval_var on each Var encountered. If evaluation errors (e.g. from a division by zero), gives up and returns the original expression. The force flag forces evaluation to proceed even if no sub-expressions changed (required if evaluating an expression that has not been constructed using smart constructors).