Module Solvers.Smt_utils

module RoundingMode : sig ... end

val float_shape : int -> int list

float_shape n is the shape of a IEEE float of a given size in bits. Returns a two element list [exp, mant], where exp is the number of exponent bits, and mant is the number of mantissa/significand bits. mant includes the hidden bit which is always 1, as per SMT-lib's expectations. Always holds that n = mant + exp. Only implemented for n = 16, 32, 64, 128.

val rm : Simple_smt.sexp

val t_f16 : Simple_smt.sexp

val t_f32 : Simple_smt.sexp

val t_f64 : Simple_smt.sexp

val t_f128 : Simple_smt.sexp

val f32_k : float -> Simple_smt.sexp

val f64_k : float -> Simple_smt.sexp

val f128_k : float -> Simple_smt.sexp

val f16_k : float -> Simple_smt.sexp

val fp_abs : Simple_smt.sexp -> Simple_smt.sexp

val fp_eq : Simple_smt.sexp -> Simple_smt.sexp -> Simple_smt.sexp

val fp_leq : Simple_smt.sexp -> Simple_smt.sexp -> Simple_smt.sexp

val fp_lt : Simple_smt.sexp -> Simple_smt.sexp -> Simple_smt.sexp

val fp_add : Simple_smt.sexp -> Simple_smt.sexp -> Simple_smt.sexp

val fp_sub : Simple_smt.sexp -> Simple_smt.sexp -> Simple_smt.sexp

val fp_mul : Simple_smt.sexp -> Simple_smt.sexp -> Simple_smt.sexp

val fp_div : Simple_smt.sexp -> Simple_smt.sexp -> Simple_smt.sexp

val fp_rem : Simple_smt.sexp -> Simple_smt.sexp -> Simple_smt.sexp

val fp_is : Stdlib.fpclass -> Simple_smt.sexp -> Simple_smt.sexp

val fp_round : RoundingMode.t -> Simple_smt.sexp -> Simple_smt.sexp

val float_of_bv : int -> Simple_smt.sexp -> Simple_smt.sexp

val float_of_ubv : RoundingMode.t -> int -> Simple_smt.sexp -> Simple_smt.sexp

val float_of_sbv : RoundingMode.t -> int -> Simple_smt.sexp -> Simple_smt.sexp

val ubv_of_float : RoundingMode.t -> int -> Simple_smt.sexp -> Simple_smt.sexp

val sbv_of_float : RoundingMode.t -> int -> Simple_smt.sexp -> Simple_smt.sexp

val int_of_bv : bool -> Simple_smt.sexp -> Simple_smt.sexp

val bv_of_int : int -> Simple_smt.sexp -> Simple_smt.sexp

val bv_nego : Simple_smt.sexp -> Simple_smt.sexp

val bv_uaddo : Simple_smt.sexp -> Simple_smt.sexp -> Simple_smt.sexp

val bv_saddo : Simple_smt.sexp -> Simple_smt.sexp -> Simple_smt.sexp

val bv_umulo : Simple_smt.sexp -> Simple_smt.sexp -> Simple_smt.sexp

val bv_smulo : Simple_smt.sexp -> Simple_smt.sexp -> Simple_smt.sexp

val reset : Simple_smt.sexp