Module Soteria.Stats

module Hstring = Soteria.Soteria_std.Hashtbl.Hstring

module Hsset = Soteria.Soteria_std.Hashset.Hstring

module Config : sig ... end

type 'range stats = {

mutable exec_time : float;
(*
Execution time for the whole execution
*)

mutable sat_time : float;
(*
Time spent answering sat queries
*)

give_up_reasons : 'range Soteria.Soteria_std.Dynarray.t Hstring.t;
(*
For each "give up reason" (a string) we keep track of the list of code ranges at which we reached it
*)

missing_without_fixes : string Soteria.Soteria_std.Dynarray.t;
(*
A list of reasons why we couldn't get a fix for a given Missing error
*)

mutable branch_number : int;
(*
Number of explored branches. Note that this includes the number of attempted branches that ended up vanishing: it may be greated than the length of the list obtained at the end of execution.
*)

mutable steps_number : int;
(*
Number of steps taken
*)

mutable unexplored_branch_number : int;
(*
Number of branches that were not explored because of fuel exhaustion. This includes branches not taking due to missing branch fuel, and branches that were cut due to step fuel.
*)

mutable sat_checks : int;
(*
Number of times the SAT solver was called
*)

mutable sat_unknowns : int;
(*
Number of times the SAT solver returned unknown
*)

}

val stats_to_yojson : 'range. ('range -> Yojson.Safe.t) -> 'range stats -> Yojson.Safe.t

val stats_of_yojson : 'range. (Yojson.Safe.t -> 'range Ppx_deriving_yojson_runtime.error_or) -> Yojson.Safe.t -> 'range stats Ppx_deriving_yojson_runtime.error_or

val _ : (Yojson.Safe.t -> 'a Ppx_deriving_yojson_runtime.error_or) -> Yojson.Safe.t -> 'a stats Ppx_deriving_yojson_runtime.error_or

type ('a, 'range) with_stats = {

res : 'a;

stats : 'range stats;

}

val map_with_stats : ('a -> 'b) -> ('a, 'c) with_stats -> ('b, 'c) with_stats

val create : unit -> 'a stats

val with_empty_stats : 'a -> ('a, 'b) with_stats

module type CodeRange = sig ... end

module type S = sig ... end

module Make (Range : CodeRange) : S with module Range = Range