Module Sym_states.Pure_fun

State model component of pure functions. A piece of state represents a function `f()` that returns a value in a certain set. It is quite similar to the agreement algebra, except that there is no allocation, there necessarily exists a value, and if we don't know it, we instantiate it on read/consumption.

module type Codom = sig ... end

module Make (C : Codom) : sig ... end