Module Tree_block.MemVal

The input module of Tree_block. A memory value t represents an owned part of the tree block, with the property that it can be split or merged as needed.

The merge doesn't need to preserve all information. For instance consider the memory values Value | Uninit | PartlyUninit; merging nodes Value and Uninit may yield a PartlyUninit node, and that node will contain both children, ensuring no information is lost.

Furthermore, to enable bi-abduction, a memory value can potentially be serialized into one or more predicates (named serialized). Consuming should extract that predicate from the given tree; producing adds it onto it.

serialized mustn't store information about the offset or length it applies to, as Tree_block wraps it into a structure containing this information.

Parameters

module Symex : Symex.Base

Signature

module type S = sig ... end